Renamed reportscollector package to collector
